pyproject.toml
[project]
name = "hebrew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "uvicorn>=0.40.0",
]
